+++ opentelemetry-semantic-conventions/src/resource.rs
@@ -9,7 +9,7 @@
 //! The [resource semantic conventions] define a set of standardized attributes
 //! to be used in `Resource`s.
 //!
-//! [resource semantic conventions]: https://github.com/open-telemetry/opentelemetry-specification/tree/master/specification/resource/semantic_conventions
+//! [resource semantic conventions]: https://github.com/open-telemetry/semantic-conventions/tree/main/model/resource
 //!
 //! ## Usage
 //!
@@ -99,7 +99,7 @@ pub const CLOUD_REGION: Key = Key::from_static_str("cloud.region");
 /// * **AWS Lambda:** The function [ARN](https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html).
 ///   Take care not to use the &#34;invoked ARN&#34; directly but replace any
 ///   [alias suffix](https://docs.aws.amazon.com/lambda/latest/dg/configuration-aliases.html)
-///   with the resolved function version, as the same runtime instance may be invocable with
+///   with the resolved function version, as the same runtime instance may be invokable with
 ///   multiple different aliases.
 /// * **GCP:** The [URI of the resource](https://cloud.google.com/iam/docs/full-resource-names)
 /// * **Azure:** The [Fully Qualified Resource ID](https://docs.microsoft.com/en-us/rest/api/resources/resources/get-by-id) of the invoked function,