+++ opentelemetry-appender-tracing/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## vNext
 
+### Changed
+
+- Bump MSRV to 1.65 [#1318](https://github.com/open-telemetry/opentelemetry-rust/pull/1318)
+
+### Added
+
 - Add log appender versions to loggers (#1182)
 
 ## v0.1.0