+++ opentelemetry-datadog/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Bump MSRV to 1.65 [#1318](https://github.com/open-telemetry/opentelemetry-rust/pull/1318)
 - Bump MSRV to 1.64 [#1203](https://github.com/open-telemetry/opentelemetry-rust/pull/1203)
 
 ### Fixed
@@ -22,7 +23,7 @@
 - Fix the array encoding length of datadog version 05 exporter #1002
 
 ## v0.7.0
-  
+
 ### Added
 - [Breaking] Add support for unified tagging [#931](https://github.com/open-telemetry/opentelemetry-rust/pull/931).